+++ gulpfile.js
@@ -14,6 +14,47 @@ var fs = require("fs"),
   mdox = require("gulp-mdox");
 
 // ----------------------------------------------------------------------------
+// Globals
+// ----------------------------------------------------------------------------
+// Sauce labs environments.
+var SAUCE_ENVS = {
+  /*jshint camelcase:false */
+  // Already tested in Travis.
+  // sl_firefox: {
+  //   base: "SauceLabs",
+  //   browserName: "firefox"
+  // },
+  sl_chrome: {
+    base: "SauceLabs",
+    browserName: "chrome"
+  },
+  sl_safari: {
+    base: "SauceLabs",
+    browserName: "safari",
+    platform: "OS X 10.9"
+  },
+  sl_ie_9: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "9"
+  },
+  sl_ie_10: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "10"
+  },
+  sl_ie_11: {
+    base: "SauceLabs",
+    browserName: "internet explorer",
+    platform: "Windows 7",
+    version: "11"
+  }
+  /*jshint camelcase:true */
+};
+
+// ----------------------------------------------------------------------------
 // Helpers
 // ----------------------------------------------------------------------------
 // Strip comments from JsHint JSON files (naive).
@@ -80,7 +121,7 @@ var testFrontend = function (opts) {
       .pipe(karma(_.extend({
         frameworks: ["mocha"],
         port: 9999,
-        reporters: "mocha",
+        reporters: ["mocha"],
         client: {
           mocha: {
             ui: "bdd"
@@ -101,6 +142,20 @@ gulp.task("test:frontend:dev", testFrontend({
 gulp.task("test:frontend:ci", testFrontend({
   singleRun: true,
   browsers: ["PhantomJS", "Firefox"]
+}));
+
+gulp.task("test:frontend:sauce", testFrontend({
+  singleRun: true,
+  reporters: ["mocha", "saucelabs"],
+  sauceLabs: {
+    testName: "chai-jq - Frontend Unit Tests",
+    public: "public"
+  },
+  // Timeouts: Allow "n" minutes before saying "good enough". See also:
+  // https://github.com/angular/angular.js/blob/master/karma-shared.conf.js
+  captureTimeout: 0, // Pass through to SL.
+  customLaunchers: SAUCE_ENVS,
+  browsers: Object.keys(SAUCE_ENVS)
 }));
 
 gulp.task("test:frontend:all", testFrontend({